Groundwork for image map functionality
The image map tag that is generated by Circos is being added to the page
but hovering over the image does not use the image map. In addition,
only one area tag is created for the map when the entire plot may
actually have two or three tracks in total.